+++ packages/core/index.js
@@ -1,6 +1,1 @@
-export kind from './kind';
-export hoc from './hoc';
-
-export default {
-	version: '1.0.0-beta.3'
-};
+// This space intentionally left blank