+++ packages/spotlight/src/spotlight.js
@@ -7,10 +7,22 @@
  * Licensed under the MPL license.
  */
 
-import Accelerator from '@enact/core/Accelerator';
+/**
+ * Exports the {@link spotlight.Spotlight} object used for controlling spotlight behavior and the
+ * {@link spotlight.Spotlight.getDirection} function for mapping a keycode to a spotlight direction.
+ *
+ * The default export is {@link spotlight.Spotlight}.
+ *
+ * @module spotlight
+ */
+
 import {is} from '@enact/core/keymap';
-import {startJob} from '@enact/core/jobs';
-import {spottableClass} from './spottable';
+import {Job} from '@enact/core/util';
+import last from 'ramda/src/last';
+
+import Accelerator from '../Accelerator';
+import {spotlightRootContainerName} from '../SpotlightRootDecorator';
+import {spottableClass} from '../Spottable';
 
 const isDown = is('down');
 const isEnter = is('enter');
@@ -22,9 +34,9 @@ const isUp = is('up');
  * Translates keyCodes into 5-way direction descriptions (e.g. `'down'`)
  *
  * @function
- * @memberof spotlight
+ * @memberof spotlight.Spotlight
  * @param {Number} keyCode - Key code to analyze
- * @returns {String|false} - One of `'up'`, `'down'`, `'left'`, `'right'` or false if not a direction key
+ * @returns {String|false} - One of `'up'`, `'down'`, `'left'`, `'right'` or `false` if not a direction key
  * @public
  */
 const getDirection = function (keyCode) {
@@ -37,7 +49,6 @@ const isPointerEvent = (target) => ('x' in target && 'y' in target);
 const isPointerShow = is('pointerShow');
 const isPointerHide = is('pointerHide');
 
-const spotlightRootContainerName = 'spotlightRootDecorator';
 const SpotlightAccelerator = new Accelerator();
 
 /**
@@ -110,15 +121,6 @@ const Spotlight = (function () {
 	 * @default true
 	 */
 	let _pointerMode = true;
-
-	/*
-	 * Length of time in milliseconds required after hiding pointer before 5-way keys
-	 * are processed.
-	 *
-	 * @type {Number}
-	 * @default 30
-	 */
-	let _pointerHiddenToKeyTimeout = 30;
 
 	/*
 	* polyfills
@@ -605,10 +607,27 @@ const Spotlight = (function () {
 
 		for (let i = containers.length; i-- > 0;) {
 			const id = containers[i];
-			if (!_containers.get(id).selectorDisabled && isNavigable(elem, id, true)) {
+			const config = _containers.get(id);
+			if (!config.selectorDisabled && matchSelector(elem, config.selector)) {
 				return id;
 			}
 		}
+	}
+
+	// returns an array of ids for containers that wrap the element, in order of outer-to-inner, with
+	// the last array item being the immediate container id of the element.
+	function getContainerIds (elem) {
+		const containerIds = [..._containers.keys()];
+		const matches = [];
+
+		for (let i = 0, containers = containerIds.length; i < containers; ++i) {
+			const id = containerIds[i];
+			const config = _containers.get(id);
+			if (!config.selectorDisabled && matchSelector(elem, config.selector)) {
+				matches.push(id);
+			}
+		}
+		return matches;
 	}
 
 	function getContainerNavigableElements (containerId) {
@@ -639,13 +658,19 @@ const Spotlight = (function () {
 		return lastFocusedElement;
 	}
 
-	function focusElement (elem, containerId, fromPointer) {
+	function setContainerLastFocusedElement (elem, containerIds) {
+		for (let i = 0, containers = containerIds.length; i < containers; ++i) {
+			_containers.get(containerIds[i]).lastFocusedElement = elem;
+		}
+	}
+
+	function focusElement (elem, containerIds, fromPointer) {
 		if (!elem) {
 			return false;
 		}
 
 		if ((_pointerMode && !fromPointer)) {
-			_containers.get(containerId).lastFocusedElement = elem;
+			setContainerLastFocusedElement(elem, containerIds);
 			return false;
 		}
 
@@ -656,7 +681,7 @@ const Spotlight = (function () {
 				currentFocusedElement.blur();
 			}
 			elem.focus();
-			focusChanged(elem, containerId);
+			focusChanged(elem, containerIds);
 		};
 
 		if (_duringFocusChange) {
@@ -680,16 +705,17 @@ const Spotlight = (function () {
 
 		_duringFocusChange = false;
 
-		focusChanged(elem, containerId);
+		focusChanged(elem, containerIds);
 		return true;
 	}
 
-	function focusChanged (elem, containerId) {
-		if (!containerId) {
-			containerId = getContainerId(elem);
+	function focusChanged (elem, containerIds) {
+		if (!containerIds || !containerIds.length) {
+			containerIds = getContainerIds(elem);
 		}
+		const containerId = last(containerIds);
 		if (containerId) {
-			_containers.get(containerId).lastFocusedElement = elem;
+			setContainerLastFocusedElement(elem, containerIds);
 			_lastContainerId = containerId;
 		}
 	}
@@ -705,9 +731,9 @@ const Spotlight = (function () {
 		} else {
 			let next = parseSelector(selector)[0];
 			if (next) {
-				let nextContainerId = getContainerId(next);
-				if (isNavigable(next, nextContainerId)) {
-					return focusElement(next, nextContainerId);
+				const nextContainerIds = getContainerIds(next);
+				if (isNavigable(next, last(nextContainerIds))) {
+					return focusElement(next, nextContainerIds);
 				}
 			}
 		}
@@ -747,7 +773,7 @@ const Spotlight = (function () {
 			}
 
 			if (next) {
-				return focusElement(next, id);
+				return focusElement(next, range);
 			}
 		}
 
@@ -767,9 +793,9 @@ const Spotlight = (function () {
 				return focusExtendedSelector(next);
 			}
 
-			let nextContainerId = getContainerId(next);
-			if (isNavigable(next, nextContainerId)) {
-				return focusElement(next, nextContainerId);
+			const nextContainerIds = getContainerIds(next);
+			if (isNavigable(next, last(nextContainerIds))) {
+				return focusElement(next, nextContainerIds);
 			}
 		}
 		return false;
@@ -786,8 +812,9 @@ const Spotlight = (function () {
 		return {allNavigableElements, containerNavigableElements};
 	}
 
-	function focusNext (next, direction, currentContainerId) {
-		const nextContainerId = getContainerId(next);
+	function focusNext (next, direction, currentContainerId, currentFocusedElement) {
+		const nextContainerIds = getContainerIds(next);
+		const nextContainerId = last(nextContainerIds);
 
 		if (currentContainerId !== nextContainerId) {
 			if (_5WayKeyHold) {
@@ -801,20 +828,22 @@ const Spotlight = (function () {
 			}
 
 			let enterToElement;
-			switch (_containers.get(nextContainerId).enterTo) {
-				case 'last-focused':
-					enterToElement = getContainerLastFocusedElement(nextContainerId) || getContainerDefaultElement(nextContainerId);
-					break;
-				case 'default-element':
-					enterToElement = getContainerDefaultElement(nextContainerId);
-					break;
+			if (!isNavigable(currentFocusedElement, nextContainerId, true)) {
+				switch (_containers.get(nextContainerId).enterTo) {
+					case 'last-focused':
+						enterToElement = getContainerLastFocusedElement(nextContainerId) || getContainerDefaultElement(nextContainerId);
+						break;
+					case 'default-element':
+						enterToElement = getContainerDefaultElement(nextContainerId);
+						break;
+				}
 			}
 			if (enterToElement) {
 				next = enterToElement;
 			}
 		}
 
-		return focusElement(next, nextContainerId);
+		return focusElement(next, nextContainerIds);
 	}
 
 	function spotNextFromPoint (direction, position, containerId) {
@@ -896,13 +925,24 @@ const Spotlight = (function () {
 				destination: next,
 				reverse: _reverseDirections[direction]
 			};
-			return focusNext(next, direction, currentContainerId);
+			return focusNext(next, direction, currentContainerId, currentFocusedElement);
 		} else if (gotoLeaveFor(currentContainerId, direction)) {
 			return true;
 		}
 
 		return false;
 	}
+
+	// 30ms (_pointerHiddenToKeyTimeout) is semi-arbitrary, to account for the time it takes for the
+	// following directional key event to fire, and to prevent momentary spotting of the last
+	// focused item - needs to be a value large enough to account for the potentially-trailing
+	// event, but not too large that another unrelated event can be fired inside the window
+	const hidePointerJob = new Job(function () {
+		_pointerMode = false;
+		if (!getCurrent() && _lastContainerId) {
+			Spotlight.focus(getContainerLastFocusedElement(_lastContainerId));
+		}
+	}, 30);
 
 	function preventDefault (evt) {
 		evt.preventDefault();
@@ -924,13 +964,14 @@ const Spotlight = (function () {
 			}
 		}
 
-		const currentContainerId = getContainerId(currentFocusedElement);
+		const currentContainerIds = getContainerIds(currentFocusedElement);
+		const currentContainerId = last(currentContainerIds);
 		if (!currentContainerId) {
 			return;
 		}
 
 		if (direction && !spotNext(direction, currentFocusedElement, currentContainerId) && currentFocusedElement !== document.activeElement) {
-			focusElement(currentFocusedElement, currentContainerId);
+			focusElement(currentFocusedElement, currentContainerIds);
 		}
 	}
 
@@ -965,17 +1006,7 @@ const Spotlight = (function () {
 		}
 
 		if (isPointerHide(keyCode)) {
-			// 30ms (_pointerHiddenToKeyTimeout) is semi-arbitrary, to account for the time it
-			// takes for the following directional key event to fire, and to prevent momentary
-			// spotting of the last focused item - needs to be a value large enough to account
-			// for the potentially-trailing event, but not too large that another unrelated
-			// event can be fired inside the window
-			startJob('hidePointer', () => {
-				_pointerMode = false;
-				if (!getCurrent() && _lastContainerId) {
-					Spotlight.focus(getContainerLastFocusedElement(_lastContainerId));
-				}
-			}, _pointerHiddenToKeyTimeout);
+			hidePointerJob.start();
 		} else if (isPointerShow(keyCode)) {
 			_pointerMode = true;
 		} else {
@@ -1005,14 +1036,16 @@ const Spotlight = (function () {
 		const target = getNavigableTarget(evt.target); // account for child controls
 
 		if (target && target !== getCurrent()) { // moving over a focusable element
-			focusElement(target, getContainerId(target), true);
+			focusElement(target, getContainerIds(target), true);
 			preventDefault(evt);
 		}
 	}
 
 	function onMouseMove (evt) {
+		const pointerMode = _pointerMode;
+
 		// Chrome emits mousemove on scroll, but client coordinates do not change.
-		if (!_pointerMode && (evt.clientX === _pointerX) && (evt.clientY === _pointerY)) {
+		if (!pointerMode && (evt.clientX === _pointerX) && (evt.clientY === _pointerY)) {
 			return;
 		}
 
@@ -1027,9 +1060,23 @@ const Spotlight = (function () {
 		}
 
 		const current = getCurrent();
+		const currentContainsTarget = current ? current.contains(evt.target) : false;
 
-		if (current && !getNavigableTarget(evt.target)) { // we are moving over a non-focusable element, so we force a blur to occur
+		// calling `getNavigableTarget()` is a heavy operation during `mousemove`, so we specifically guard
+		// against unnecessarily executing it
+		if (pointerMode && current && !currentContainsTarget) {
+			// we are moving over a non-focusable element, so we force a blur to occur
 			current.blur();
+		} else if (!pointerMode && !(current && currentContainsTarget)) {
+			const target = getNavigableTarget(evt.target);
+
+			if (!target && current) {
+				// we are moving over a non-focusable element, so we force a blur to occur
+				current.blur();
+			} else if (target && (!current || target !== current)) {
+				// we are moving over a focusable element, so we set focus to the target
+				focusElement(target, getContainerIds(target), true);
+			}
 		}
 	}
 
@@ -1054,7 +1101,8 @@ const Spotlight = (function () {
 	 */
 	const exports = /** @lends spotlight.Spotlight.prototype */ { // eslint-disable-line no-shadow
 		/**
-		 * Initializes Spotlight. This is generally handled by {@link spotlight.SpotlightRootDecorator}.
+		 * Initializes Spotlight. This is generally handled by
+		 * {@link spotlight/SpotlightRootDecorator.SpotlightRootDecorator}.
 		 *
 		 * @memberof spotlight.Spotlight.prototype
 		 * @public
@@ -1286,9 +1334,10 @@ const Spotlight = (function () {
 					result = focusExtendedSelector(elem);
 				}
 			} else {
-				let nextContainerId = getContainerId(elem);
+				const nextContainerIds = getContainerIds(elem);
+				const nextContainerId = last(nextContainerIds);
 				if (isNavigable(elem, nextContainerId)) {
-					result = focusElement(elem, nextContainerId);
+					result = focusElement(elem, nextContainerIds);
 				}
 			}
 
@@ -1442,4 +1491,7 @@ const Spotlight = (function () {
 })();
 
 export default Spotlight;
-export {Spotlight, spotlightRootContainerName, getDirection};
+export {
+	getDirection,
+	Spotlight
+};