+++ packages/spotlight/CHANGELOG.md
@@ -6,7 +6,22 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
-- `spotlight.Spottable` behavior for forwarding selection `onKeyUp` events
+- `spotlight/Spottable` behavior for forwarding selection `onKeyUp` events
+- `spotlight/Spotlight` behavior to properly save the last-focused element for nested containers
+
+### Removed
+
+- `spotlight.Spottable` replaced by `spotlight/Spottable`
+- `spotlight.spottableClass` replaced by `spotlight/Spottable.spottableClass`
+- `spotlight.SpotlightContainerDecorator` replaced by `spotlight/SpotlightContainerDecorator`
+- `spotlight.spotlightDefaultClass` replaced by `spotlight/SpotlightContainerDecorator.spotlightDefaultClass`
+- `spotlight.SpotlightRootDecorator` replaced by `spotlight/SpotlightRootDecorator`
+
+## [1.0.0-beta.4] - 2017-03-10
+
+### Fixed
+
+- `spotlight.Spotlight` pointer behavior where upon immediately entering pointer-mode, hovering over a spottable component may result in the component not receiving focus
 
 ### Changed